Move mails worker in separate package
This is done to prevent import cycles
